backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pdfplumber==0.10.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
Werkzeug==3.0.1
numpy==1.24.3
scikit-learn==1.3.0 